olson working, back not yet

+++ source2_new/SuOlsonAll.h
@@ -1,11 +1,11 @@
 #ifndef SuOlsonAll_H_
 #define SuOlsonAll_H_
-float calculateWeff(int,int);
-float calculateWeffNonAvg(int,int);
-float getdx();
-float getdt();
-float calculateB(float);
-float calculateMu2(float weff);
-float getOpacity(int space,int time1);
-float getFinc();
+double calculateWeff(int,int);
+double calculateWeffNonAvg(int,int);
+double getdx();
+double getdt();
+double calculateB(double);
+double calculateMu2(double weff);
+double getOpacity(int space,int time1);
+double getFinc();
 #endif
